Add Publishing Apps index config
In preparation to index content based on its publishing app, we add config to
distinguish between apps whose content should be indexed by Rummager, and apps
whose content should not be indexed.  There are some special cases where only
some of the formats that an app can publish are indexable, and also certain
routes that can be indexed.  We also add a module which allows the app to
easily check if a content item is indexable or not.